+++ rgpycrumbs/eon/con_splitter.py
@@ -11,6 +11,7 @@
 #   "ase",
 #   "rich",
 #   "readcon>=0.7.0",
+#   "rgpycrumbs>=1.9.13",
 # ]
 # ///