+++ rgpycrumbs/eon/generate_nwchem_input.py
@@ -11,6 +11,7 @@
 #   "click>=8.2.1",
 #   "rich",
 #   "pychum",
+#   "rgpycrumbs>=1.9.13",
 # ]
 # ///