+++ rgpycrumbs/eon/to_mlflow.py
@@ -12,6 +12,7 @@
 #     "rich",
 #     "ase",
 #     "mlflow",
+#   "rgpycrumbs>=1.9.13",
 # ]
 # ///