+++ rgpycrumbs/plumed/direct_reconstruction.py
@@ -2,6 +2,7 @@
 # requires-python = ">=3.11"
 # dependencies = [
 #  "pandas", "numpy", "matplotlib", "cmcrameri", "chemparseplot"
+#   "rgpycrumbs>=1.9.13",
 # ]
 # ///